refactor: split schemas binding into package
