+++ pkgs/standards/autoapi/autoapi/v3/bindings/schemas/__init__.py
@@ -1,0 +1,6 @@
+# autoapi/v3/bindings/schemas/__init__.py
+from __future__ import annotations
+
+from .builder import build_and_attach
+
+__all__ = ["build_and_attach"]
